Refactor : modify package structure
